+++ memory-mcp/src/bin/server_impl/tools/memory_handlers.rs
@@ -9,7 +9,6 @@ use memory_mcp::mcp::tools::embeddings::{
     ConfigureEmbeddingsInput, EmbeddingProviderStatusInput, GenerateEmbeddingInput,
     QuerySemanticMemoryInput, SearchByEmbeddingInput,
 };
-use memory_mcp::mcp::tools::pattern_search::{RecommendPatternsInput, SearchPatternsInput};
 use memory_mcp::mcp::tools::quality_metrics::QualityMetricsInput;
 use serde_json::json;
 
@@ -417,189 +416,6 @@ pub async fn handle_embedding_provider_status(
     let input: EmbeddingProviderStatusInput = serde_json::from_value(args)?;
 
     let result = server.execute_embedding_provider_status_tool(input).await?;
-
-    let content = vec![Content::Text {
-        text: serde_json::to_string_pretty(&result)?,
-    }];
-    Ok(content)
-}
-
-/// Handle search_patterns tool
-pub async fn handle_search_patterns(
-    server: &mut MemoryMCPServer,
-    arguments: Option<Value>,
-) -> anyhow::Result<Vec<Content>> {
-    let args: Value = arguments.ok_or_else(|| anyhow::anyhow!("Missing arguments"))?;
-    let client_id = get_client_id(&args);
-    let input: SearchPatternsInput = serde_json::from_value(args)?;
-    let domain = input.domain.clone();
-
-    // Access memory through the server's memory field
-    let result = server.execute_search_patterns(input).await;
-
-    // Audit log the operation
-    let result_count = result.as_ref().map(json_value_len).unwrap_or(0);
-    server
-        .audit_logger()
-        .log_pattern_search(&client_id, &domain, result_count, result.is_ok())
-        .await;
-
-    let content = vec![Content::Text {
-        text: serde_json::to_string_pretty(&result?)?,
-    }];
-    Ok(content)
-}
-
-/// Handle recommend_patterns tool
-pub async fn handle_recommend_patterns(
-    server: &mut MemoryMCPServer,
-    arguments: Option<Value>,
-) -> anyhow::Result<Vec<Content>> {
-    let args: Value = arguments.ok_or_else(|| anyhow::anyhow!("Missing arguments"))?;
-    let client_id = get_client_id(&args);
-    let input: RecommendPatternsInput = serde_json::from_value(args)?;
-    let domain = input.domain.clone();
-
-    // Access memory through the server's memory field
-    let result = server.execute_recommend_patterns(input).await;
-
-    // Audit log the operation
-    let recommendation_count = result.as_ref().map(json_value_len).unwrap_or(0);
-    server
-        .audit_logger()
-        .log_recommend_patterns(&client_id, &domain, recommendation_count, result.is_ok())
-        .await;
-
-    let content = vec![Content::Text {
-        text: serde_json::to_string_pretty(&result?)?,
-    }];
-    Ok(content)
-}
-
-/// Handle recommend_playbook tool (ADR-044 Feature 1)
-pub async fn handle_recommend_playbook(
-    server: &mut MemoryMCPServer,
-    arguments: Option<Value>,
-) -> anyhow::Result<Vec<Content>> {
-    let args: Value = arguments.ok_or_else(|| anyhow::anyhow!("Missing arguments"))?;
-    let client_id = get_client_id(&args);
-
-    let task_description = args
-        .get("task_description")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| anyhow::anyhow!("Missing 'task_description' parameter"))?
-        .to_string();
-
-    let domain = args
-        .get("domain")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| anyhow::anyhow!("Missing 'domain' parameter"))?
-        .to_string();
-
-    let task_type_str = args
-        .get("task_type")
-        .and_then(|v| v.as_str())
-        .unwrap_or("code_generation");
-
-    let task_type = match task_type_str {
-        "code_generation" => memory_core::TaskType::CodeGeneration,
-        "debugging" => memory_core::TaskType::Debugging,
-        "refactoring" => memory_core::TaskType::Refactoring,
-        "testing" => memory_core::TaskType::Testing,
-        "analysis" => memory_core::TaskType::Analysis,
-        "documentation" => memory_core::TaskType::Documentation,
-        _ => memory_core::TaskType::CodeGeneration,
-    };
-
-    let max_steps = args.get("max_steps").and_then(|v| v.as_u64()).unwrap_or(5) as usize;
-
-    let language = args
-        .get("language")
-        .and_then(|v| v.as_str())
-        .map(|s| s.to_string());
-    let framework = args
-        .get("framework")
-        .and_then(|v| v.as_str())
-        .map(|s| s.to_string());
-
-    let tags = args
-        .get("tags")
-        .and_then(|v| v.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let context = memory_core::TaskContext {
-        domain: domain.clone(),
-        language,
-        framework,
-        complexity: memory_core::ComplexityLevel::Moderate,
-        tags,
-    };
-
-    // Retrieve playbooks from memory
-    let playbooks = server
-        .memory()
-        .retrieve_playbooks(
-            &task_description,
-            &domain,
-            task_type,
-            context,
-            1, // max_playbooks
-            max_steps,
-        )
-        .await;
-
-    let result = serde_json::to_value(&playbooks)?;
-
-    // Audit log the operation
-    server
-        .audit_logger()
-        .log_playbook_recommendation(&client_id, &task_description, playbooks.len(), true)
-        .await;
-
-    let content = vec![Content::Text {
-        text: serde_json::to_string_pretty(&result)?,
-    }];
-    Ok(content)
-}
-
-/// Handle explain_pattern tool (ADR-044 Feature 1)
-pub async fn handle_explain_pattern(
-    server: &mut MemoryMCPServer,
-    arguments: Option<Value>,
-) -> anyhow::Result<Vec<Content>> {
-    let args: Value = arguments.ok_or_else(|| anyhow::anyhow!("Missing arguments"))?;
-    let client_id = get_client_id(&args);
-
-    let pattern_id_str = args
-        .get("pattern_id")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| anyhow::anyhow!("Missing 'pattern_id' parameter"))?;
-
-    let pattern_id = uuid::Uuid::parse_str(pattern_id_str)
-        .map_err(|e| anyhow::anyhow!("Invalid pattern_id: {}", e))?;
-
-    // Get pattern explanation from memory
-    let explanation = server
-        .memory()
-        .explain_pattern(pattern_id)
-        .await
-        .ok_or_else(|| anyhow::anyhow!("Pattern not found: {}", pattern_id))?;
-
-    let result = serde_json::json!({
-        "pattern_id": pattern_id_str,
-        "explanation": explanation
-    });
-
-    // Audit log the operation
-    server
-        .audit_logger()
-        .log_pattern_explanation(&client_id, pattern_id_str, true)
-        .await;
 
     let content = vec![Content::Text {
         text: serde_json::to_string_pretty(&result)?,
